+++ packages/core/src/cli/config.ts
@@ -3,6 +3,7 @@ import { isAbsolute, join } from 'path';
 import {
   color,
   logger,
+  debounce,
   type RsbuildConfig as BaseRsbuildConfig,
 } from '@rsbuild/shared';
 import { restartDevServer } from '../server/restart';
@@ -51,11 +52,20 @@ const resolveConfigPath = (customConfig?: string) => {
 
 async function watchConfig(configFile: string) {
   const chokidar = await import('@rsbuild/shared/chokidar');
-  const watcher = chokidar.watch(configFile, {});
-  const callback = async () => {
-    watcher.close();
-    await restartDevServer({ filePath: configFile });
-  };
+
+  const watcher = chokidar.watch(configFile, {
+    // If watching fails due to read permissions, the errors will be suppressed silently.
+    ignorePermissionErrors: true,
+  });
+
+  const callback = debounce(
+    async () => {
+      watcher.close();
+      await restartDevServer({ filePath: configFile });
+    },
+    // set 300ms debounce to avoid restart frequently
+    300,
+  );
 
   watcher.on('change', callback);
   watcher.on('unlink', callback);
@@ -67,20 +77,25 @@ export async function loadConfig(
   const configFile = resolveConfigPath(customConfig);
 
   if (configFile) {
-    const { default: jiti } = await import('../../compiled/jiti');
-    const loadConfig = jiti(__filename, {
-      esmResolve: true,
-      // disable require cache to support restart CLI and read the new config
-      requireCache: false,
-      interopDefault: true,
-    });
+    try {
+      const { default: jiti } = await import('../../compiled/jiti');
+      const loadConfig = jiti(__filename, {
+        esmResolve: true,
+        // disable require cache to support restart CLI and read the new config
+        requireCache: false,
+        interopDefault: true,
+      });
 
-    const command = process.argv[2];
-    if (command === 'dev') {
-      watchConfig(configFile);
+      const command = process.argv[2];
+      if (command === 'dev') {
+        watchConfig(configFile);
+      }
+
+      return loadConfig(configFile);
+    } catch (err) {
+      logger.error(`Failed to load file: ${color.dim(configFile)}`);
+      throw err;
     }
-
-    return loadConfig(configFile);
   }
 
   return {};