+++ packages/shared/src/types/config/server.ts
@@ -1,3 +1,5 @@
+export type HtmlFallback = false | 'index';
+
 export interface ServerConfig {
   /**
    * Specify a port number for Dev Server to listen.
@@ -15,7 +17,8 @@ export interface ServerConfig {
    * Adds headers to all responses.
    */
   headers?: Record<string, string | string[]>;
+  htmlFallback?: HtmlFallback;
 }
 
 export type NormalizedServerConfig = ServerConfig &
-  Required<Pick<ServerConfig, 'port' | 'host'>>;
+  Required<Pick<ServerConfig, 'htmlFallback' | 'port' | 'host'>>;