+++ packages/compat/uni-builder/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @rsbuild/uni-builder
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [0f95693]
+- Updated dependencies [1afd033]
+- Updated dependencies [2b267b5]
+  - @rsbuild/webpack@0.0.28
+  - @rsbuild/shared@0.0.28
+  - @rsbuild/core@0.0.28
+  - @rsbuild/plugin-assets-retry@0.0.28
+  - @rsbuild/plugin-rem@0.0.28
+
 ## 0.0.27
 
 ### Patch Changes