+++ packages/doctor-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rsbuild/doctor-core
 
+## 0.0.14
+
+### Patch Changes
+
+- 25f1922: fix: rsbuild doctor compatible with rspack canary version
+  - @rsbuild/doctor-sdk@0.0.14
+  - @rsbuild/doctor-utils@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes